refactor(storage): move catalog↔Collection mapping into storage (Slice D)
Move the catalog-schema ⇄ proto Collection mapping out of CollectionService
into a new storage module so src/storage no longer references
crate::services::collection at all.

- new module crate::storage::metadata::collection_mapping: the
  collection_from_catalog_schema / catalog_schema_from_collection pair plus
  6 private conversion helpers (8 fns, ~546 LOC), moved verbatim
  (behaviour-neutral). The 3 fns called from outside the cluster
  (collection_from_catalog_schema, catalog_schema_from_collection,
  collection_table_identifier) are pub(crate); the rest stay private.
- it belongs in storage, not proximadb-catalog: the mapping round-trips
  through storage's own catalog_config json helpers
  (crate::storage::metadata::catalog_config), so a control-tier catalog home
  would depend up on storage (cycle). Storage is the natural owner — the
  mapping is driven from WAL recovery — and services now calls *down* into it.
- rewire all callers: storage WAL (3) + the collection service (read/load +
  tests) now reference crate::storage::metadata::collection_mapping::*.
- generate_unique_collection_id (an interleaved &self async method) correctly
  stays on CollectionService; trimmed the imports it/the cluster orphaned in
  manager.rs (test-only types re-scoped into the test module).

Clears the last storage→crate::services::collection edge (grep of
`services::collection` under src/storage is now empty), unblocking storage
engine extraction on the metadata side. IndexPort remains.

Verified: cargo check -p proximadb --tests, clippy --lib --bins -D,
server/bench/migrate bins, layering green, doc + capability validators.
