+++ docs/12-design/ROOT_CRATE_DECOMPOSITION_SLICE_D_PORT_INVERSION_2026_06_26.adoc
@@ -223,7 +223,7 @@ first so the hot edits are mechanical.
 | D0 | **Redirect-down (Mechanism A, existing crates).** Rewrite storage's `crate::compute::…DistanceMetric` / `crate::core::search::FilterExpression` / quant-types / config / hardware-caps / index-types references to the existing foundation crates. Pure path rewrites; verify each root path == leaf type. Several small mechanical PRs; collapses the edge count first. **DONE:** distance-kernel + FilterExpression (PR #370), hardware-caps + index-types (D0b, PR #377). **Mechanical phase complete** — the two remaining items are NOT clean redirects: quant-types resolves to *modality-tier* defs (Mechanism B, see D2 spec above) and config is a divergent root module. | Low
 | D1 | **Extract remaining shared-type leaves down.** `proximadb-search-types` (← `core::search` minus FilterExpression) and `proximadb-bloom` (← `core::bloom`). Same leaf recipe already proven. **DONE:** `proximadb-bloom` (PR #379); `proximadb-search-types` — the whole-file leaves `results`/`bounded_queue`/`sql_value_filter`/`json_value_serde` + the carved-out `json_comparison` module (this PR). `UnifiedSearchParams` stays in root (3 upward edges: compute `DistanceMetric`, modality `UnifiedQuantizationLevel`, query `FilterOptimizationHints`); the inline simple enums (`SearchMode`/`BlockPruneConfig`/…) remain interleaved with it and are deferred. | Low
 | D2 | **Define `proximadb-storage-ports`** — the `IndexPort` + `CollectionMetadataPort` traits. **DONE (CollectionMetadataPort):** crate created (`crates/storage/proximadb-storage-ports`, storage tier — same-tier deps allowed) with `async fn collection(&str) -> Result<Option<Collection>>` (the *measured* surface is one method; `get_collection_stats` turned out to be on `AxisManager`, not the service); `impl CollectionMetadataPort for CollectionService`. **Remaining:** `IndexPort` (leaks `AxisHybridQuery`/`AxisManagerQueryResult`/`AxisConfig`/`AxisVectorIndex` — per-type facade-vs-pre-extract decision; re-exporting root types into the port crate is a *cycle*, impossible). | Low–Med
-| D3 | **Refactor storage to consume injected ports.** `engine.rs`/compaction stop constructing `AxisManager` / calling `CollectionService` directly; take `Arc<dyn IndexPort>` / `Arc<dyn CollectionMetadataPort>` via constructor injection. Same concrete behavior, now behind the trait. **This is the hot, coordination-window phase.** **DONE (metadata side):** storage fields/params/setters in viper engine+flush, flush-coordinator, and background-flush-context retyped from `Arc<CollectionService>` to `Arc<dyn CollectionMetadataPort>` (these injection setters currently have no callers — latent infra — so this is a pure structural edge removal). **DONE (tenant resolver):** `collection_tenant_id` relocated *down* to a new foundation crate **`proximadb-tenant`** (`tenant_id_of(&Collection)`) — not "down to catalog" as first sketched, because tenant resolution is a *cross-cutting economics/isolation primitive* (storage, embedded, services, and the network gates all share it; coupling it to any one tier is wrong). Foundation tier = every layer depends *down* on the single fail-closed attribution seam, and resolution is a pure function of collection metadata so any node in an autoscaled fleet attributes identically without coordination (elasticity) and per-tenant meters reconcile (economics). `CollectionService::collection_tenant_id` is now a thin alias; storage/embedded call `proximadb_tenant::tenant_id_of` directly. **Remaining:** relocate the `collection_from_catalog_schema`/`catalog_schema_from_collection` mapping pair *down* to `proximadb-catalog` (control) to clear the last 3 storage→services refs; and the whole `IndexPort` wiring. | Med-High
+| D3 | **Refactor storage to consume injected ports.** `engine.rs`/compaction stop constructing `AxisManager` / calling `CollectionService` directly; take `Arc<dyn IndexPort>` / `Arc<dyn CollectionMetadataPort>` via constructor injection. Same concrete behavior, now behind the trait. **This is the hot, coordination-window phase.** **DONE (metadata side):** storage fields/params/setters in viper engine+flush, flush-coordinator, and background-flush-context retyped from `Arc<CollectionService>` to `Arc<dyn CollectionMetadataPort>` (these injection setters currently have no callers — latent infra — so this is a pure structural edge removal). **DONE (tenant resolver):** `collection_tenant_id` relocated *down* to a new foundation crate **`proximadb-tenant`** (`tenant_id_of(&Collection)`) — not "down to catalog" as first sketched, because tenant resolution is a *cross-cutting economics/isolation primitive* (storage, embedded, services, and the network gates all share it; coupling it to any one tier is wrong). Foundation tier = every layer depends *down* on the single fail-closed attribution seam, and resolution is a pure function of collection metadata so any node in an autoscaled fleet attributes identically without coordination (elasticity) and per-tenant meters reconcile (economics). `CollectionService::collection_tenant_id` is now a thin alias; storage/embedded call `proximadb_tenant::tenant_id_of` directly. **DONE (catalog↔Collection mapping):** the `collection_from_catalog_schema`/`catalog_schema_from_collection` pair + 6 private helpers (8 fns, ~546 LOC) moved verbatim out of `CollectionService` into a new storage module `crate::storage::metadata::collection_mapping` — *not* down to `proximadb-catalog` as first sketched, because the mapping reaches *up* into storage's own `catalog_config` json round-trip (control-tier can't depend on storage → cycle). Storage is its natural home (WAL-recovery driven, round-trips through storage-owned config); services now calls *down* into it. **This clears the last storage→`crate::services::collection` reference** (`generate_unique_collection_id`, an interleaved `&self` async method, correctly stayed behind). **Remaining:** the whole `IndexPort` wiring (the larger port). | Med-High
 | D4 | **Implement ports + wire injection.** `impl IndexPort for AxisManager`, `impl CollectionMetadataPort for CollectionService`; the composition root injects them. | Med
 | D5 | **Gate to zero.** `scripts/check_workspace_boundaries.py` + a storage-up-edge ratchet asserting `crate::{index,compute,services,query}` references under `src/storage` == 0 (core type refs already redirected in D0/D1). Engines (Slice E) now unblocked. | Low
 |===
